refactor: share the duplicated drive service layer through the drive package
Moves the OneDrive and SharePoint modules that differed only by provider
name into packages/drive, taking the differences as parameters: the key
layout, the manifest lookup, the workload name, and the index and blob
readers.

Refs #151
