+++ packages/drive/src/drive-ports.ts
@@ -24,6 +24,19 @@ export type DriveSnapshotManifest = OneDriveSnapshotManifest;
 export type DriveVersionWatermark = OneDriveVersionWatermark;
 
 /**
+ * What the shared version code reads from a file version index. The owning segment is named after
+ * what it owns in each provider (`owner_id` on OneDrive, `site_id` on SharePoint) and is never read
+ * here, so the shared functions take this view instead of either concrete index type.
+ */
+export interface DriveFileVersionIndexView {
+  readonly file_id: string;
+  readonly versions: readonly DriveFileVersionRecord[];
+}
+
+/** The two drive workloads, as they are named in progress events and default output paths. */
+export type DriveWorkload = 'onedrive' | 'sharepoint';
+
+/**
  * The connector surface the shared drive code touches. Both `OneDriveConnector` and
  * `SharePointSiteConnector` satisfy it structurally, so neither package has to adapt anything.
  */
